move pose_estimator into server
