+++ server/pose_estimator/src/Dockerfile
