+++ server/requirements.txt
@@ -7,3 +7,5 @@ MarkupSafe==2.1.3
 typing_extensions==4.7.1
 Werkzeug==2.2.3
 zipp==3.15.0
+matplotlib
+opencv-python
